Improve javadoc and refactor codegen package structure
Package name should be written in singularis. For a long time, codegen
have broken this convention. There was also some package names that were
confusing, like "interfaces". These have now been renamed to follow the
general trend in the other packages. Traits are for an example named
"HasXXX" and are located in a dedicated package named "trait".