generator-core: Change structure of TranslatorManager to allow overrides
This doesn't change the API, but will make it easier to continue internal development. There are now two implementations of TranslatorManager in the module. One, called "DefaultTranslatorManager" that is used by the injector by default, and one "AbstractTranslatorManager" that users can extend from other modules if they want to change the behaviour of the manager. All the logic has been moved to internal class TranslatorManagerHelper to improve code sharing between the two.