default dbmsTypeName, always start with new file
DBMS always needs a typeName to be able to load the json file, this has been set to MySQL. Also, the dbms port has been set to the default MySQL port of 3306.

The save method has been extended to check if the config file is already present. If it is, it is first removed. The init always starts with a new file.