generating the sources adds the new source folder if needed
When the sources are generating, the generator mojo now verifies if the package location is already registered as a source location. If this is not the case, it gets added to the list of registered source locations.
This removes the need for the maven-helper-plugin to additionally add the new source folder.